% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.R
\name{author_scopus}
\alias{author_scopus}
\title{Find author ids from SCOPUS}
\usage{
author_scopus(threshold = 0.9, update = FALSE)
}
\arguments{
\item{threshold}{Threshold for similarity}

\item{update}{whether update tiddlywiki}
}
\value{
A data.frame with new author id
}
\description{
Find author ids from SCOPUS
}
