% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese.R
\name{convert_pinyin}
\alias{convert_pinyin}
\title{Convert tiddlers with Chinese title into pinyin}
\usage{
convert_pinyin()
}
\value{
No return values
}
\description{
Convert tiddlers with Chinese title into pinyin
}
\details{
The converted pinyin is merged into field `aka`. An extra field `py-checked` is added to indicate
this tiddler is checked and will skip in the future.
}
