% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{get_dois}
\alias{get_dois}
\title{Get DOIs from tiddlywiki}
\usage{
get_dois(filter = NULL)
}
\arguments{
\item{filter}{filter to get dois}
}
\value{
A data frame with two columns title and doi
}
\description{
Get DOIs from tiddlywiki
}
