% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{reference}
\alias{reference}
\title{Get reference list for publications from crossref}
\usage{
reference()
}
\value{
no return
}
\description{
Get reference list for publications from crossref
}
\details{
The reference list is obtained from crossref. The field `reference` in Tiddlywiki is updated
to list all reference in the Tiddlywiki. Two fields `reference-count` and `cited-count` is also updated
for number of reference and citation for a publication, respectively.
}
