% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_outfiles}
\alias{remove_outfiles}
\title{Remove old intermediate files by dates}
\usage{
remove_outfiles(files, expired_days = NULL)
}
\arguments{
\item{files}{a list of files}

\item{expired_days}{days to expire}
}
\value{
no return
}
\description{
Intermediate files are expired by number of days (defined by option `file_expired`, 90 days in default).
As the limitation of web API, only a few number of files are removed (define by option `file_remove_max`, 3 in default).
}
