% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.R
\name{scopus_works}
\alias{scopus_works}
\title{Get works from Scopus using Scopus API.}
\usage{
scopus_works(is_new = FALSE)
}
\arguments{
\item{is_new}{Logical. Whether to only process new records.}
}
\value{
A data frame for all works obtained from Scopus
}
\description{
R package `rscopus` is used to retrieve works from Scopus according to `scopus`
field in Tiddlywiki.
}
\details{
The retrieved works are stored into subfolder `scopus` in the output folder which
is defined by option `output`.

The works will be re-downloaded after 90 days (modifying through option file_expired)
if `is_new` equals to `FALSE`
}
