% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/works.R
\name{works_all}
\alias{works_all}
\title{Get works from all resources including ORCID, Scopus}
\usage{
works_all(is_new = TRUE)
}
\arguments{
\item{is_new}{Logical. Whether to only process new records.}
}
\value{
No return
}
\description{
Get works from all resources including ORCID, Scopus
}
