% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{works_authoring}
\alias{works_authoring}
\title{Title Add authors for works}
\usage{
works_authoring(works, is_new = FALSE)
}
\arguments{
\item{works}{a data frame for works}

\item{is_new}{whether only to process new items}
}
\value{
No return
}
\description{
Tiddlers with tag bibtex-entry will be updated for new authors.
}
