% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homepage.R
\name{works_homepage}
\alias{works_homepage}
\title{Get works from homepage}
\usage{
works_homepage(is_new = FALSE)
}
\arguments{
\item{is_new}{Logical. Whether to only process new records.}
}
\value{
A data frame for all works obtained from homepage
}
\description{
Get works from homepage
}
\details{
It assumes all dois in the homepage are colleague's publication. The field `url` is to
define the homepage.

The retrieved works are stored into subfolder `homepage` in the output folder which
is defined by option `output`.

The works will be re-downloaded after 90 days (modifying through option file_expired)
if `is_new` equals to `FALSE`
}
