% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid.R
\name{works_orcid}
\alias{works_orcid}
\title{Get works from ORCID using ORCID API.}
\usage{
works_orcid(is_new = FALSE)
}
\arguments{
\item{is_new}{Logical. Whether to only process new records.}
}
\value{
A data frame for all works obtained from ORCID
}
\description{
Get works from ORCID using ORCID API.
}
\details{
R package `rorcid` is used to retrieve works from ORCID according to `orcid`
field in Tiddlywiki. Package `rorcid` is archived now and there is no alternative
R package to access ORCID API in R. This package use two system variables to
authenticate orcid (i.e. `ORCID_CLIENT_ID` and `ORCID_CLIENT_SECRET`).

The retrieved works are stored into subfolder `orcid` in the output folder which
is defined by option `output`.

The works will be re-downloaded after 90 days (modifying through option file_expired)
if `is_new` equals to `FALSE`
}
