% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researcherid.R
\name{works_wos}
\alias{works_wos}
\title{Get works from ResearchID using ResearchID API.}
\usage{
works_wos(is_new = FALSE)
}
\arguments{
\item{is_new}{Logical. Whether to only process new records.}
}
\value{
A data frame for all works obtained from WOS
}
\description{
Get works from ResearchID using ResearchID API.
}
\details{
[WOS Starter API] (https://developer.clarivate.com/apis/wos-starter) is used to retrieve works
from Web of Scient according to `researcherid` field in Tiddlywiki.
This package use environment variable to authenticate WOS (i.e. `WOS_STARTER_KEY`).

The retrieved works are stored into subfolder `researcherid` in the output folder which
is defined by option `output`.

The works will be re-downloaded after 90 days (modifying through option file_expired)
if `is_new` equals to `FALSE`
}
