% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{reference}
\alias{reference}
\title{Get reference list for publications from crossref and opencitations}
\usage{
reference(remove_old = TRUE, all = FALSE)
}
\arguments{
\item{remove_old}{whether to remove old files}

\item{all}{whether to get all reference}
}
\value{
no return
}
\description{
Get reference list for publications from crossref and opencitations
}
\details{
The reference list is obtained from crossref and opencitations. The field `reference` in Tiddlywiki is updated
to list all reference in the Tiddlywiki. Two fields `reference-count` and `cited-count` is also updated
for number of reference and citation for a publication, respectively.
}
